% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{enframe_markers}
\alias{enframe_markers}
\title{Enframe Cluster Markers}
\usage{
enframe_markers(tbl)
}
\arguments{
\item{tbl}{a tibble of marker genes}
}
\value{
a pivoted tibble of marker genes
}
\description{
Enframe Cluster Markers
}
