% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevreulPlot-package.R
\docType{package}
\name{chevreulPlot-package}
\alias{chevreulPlot}
\alias{chevreulPlot-package}
\title{chevreulPlot: Plots used in the chevreulPlot package}
\description{
Tools for plotting SingleCellExperiment objects in the chevreulPlot package. Includes functions for analysis and visualization of single-cell data. Supported by NIH grants R01CA137124 and R01EY026661 to David Cobrinik.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whtns/chevreulPlot}
  \item \url{https://whtns.github.io/chevreulPlot/}
  \item Report bugs at \url{https://github.com/cobriniklab/chevreulPlot/issues}
}

}
\author{
\strong{Maintainer}: Kevin Stachelek \email{kevin.stachelek@gmail.com} (\href{https://orcid.org/0000-0003-2085-695X}{ORCID})

Authors:
\itemize{
  \item Bhavana Bhat \email{bhavanabhat29@gmail.com}
}

}
\keyword{internal}
